pyproject.toml
[project]
name = "tds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "pandas>=2.2.3",
  "pygithub>=2.4.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3",
  "rich>=13.9.3",
  "scipy>=1.14.1",
]
[tool.ruff.lint]
select = [
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]

[dependency-groups]
dev = ["pandas-stubs>=2.2.3.241009"]
